hms_erp_project/requirements.txt
fastapi
uvicorn 
python-decouple
sqlalchemy
psycopg2 >= "2.9.6"
psycopg2-binary >= "2.9.6"
pydantic
alembic
email-validator
bcrypt
python-jose
passlib
pyjwt
pytest
httpx
pydantic-settings
asyncpg >= "0.25.0"
databases >= "0.5.3"
asyncclick
